pyproject.toml
[project]
name = "ai-digest-bot"
version = "0.1.0"
description = "Telegram bot that delivers daily AI digests filtered by user's business projects, with Notion-synced project profiles"
requires-python = ">=3.11"
authors = [{ name = "dowa13" }]
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "google-genai>=0.3.0",
    "groq>=0.11.0",
    "supabase>=2.5.0",
    "notion-client>=2.2.1",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "selectolax>=0.3.21",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
# RUF001/2/3 catch ambiguous Cyrillic chars — false positives in a Russian-language app.
# SIM105: contextlib.suppress in async hot path is uglier than try/except/pass.
ignore = ["E501", "RUF001", "RUF002", "RUF003", "SIM105"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
files = ["src/shared"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
